Allow dart/sass path to be configured, closes #15
By default the sass snapshot (or executable) and optionally,
depending on the operating system, the Dart VM executable
will be automatically downloaded and installed inside the
`_build` directory of the current app. The paths are now
configurable via the global app configuration.

Additionally we fallback if Mix is not available, such as
inside releases.